Adds empty dashboard page (#9)
* Adds empty dashboard page

The dashboard page is the first page that customers see when they click Anomaly detection in the side navigation.  If there is at least one detector, we show the detailed dashboard page. Otherwise, we show an empty dashboard page. This CR adds the empty dashboard page. 

This CR also adds logic for 
- redirecting to a detailed or empty dashboard page.
- dealing with loading status and refactoring

Testing done:
- manual testing to verify the workflow
- Add snapshot unit test for the empty dashboard page.
- Add unit test for redirect logic